railway_bea_generator/requirements.txt
pandas==2.0.3
numpy==1.24.4
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
